fix on search criteria

+++ src/main/java/egd/covid/persistence/repository/CovidEntityRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import egd.covid.persistence.entity.CovidEntity;
 
 @Repository
-public interface CovidEntityRepository extends JpaRepository<CovidEntity, Integer> {
+public interface CovidEntityRepository extends JpaRepository<CovidEntity, Long> {
 	@Query(value = "SELECT coalesce(max(id), 0) FROM CovidEntity")
 	public long getMaxId();
 }
